requirements.txt
openai == 1.108.2
pyyaml == 6.0.2
